day5/day5-1/requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
cbor==1.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dashscope==1.25.12
dataclasses-json==0.6.7
datasets==4.5.0
dill==0.4.0
distro==1.9.0
faiss-cpu==1.13.2
filelock==3.24.3
FlagEmbedding==1.3.5
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ijson==3.4.0.post0
inscriptis==2.7.0
ir_datasets==0.5.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-classic==1.0.1
langchain-community==0.2.16
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.1.147
lxml==5.4.0
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
modelscope==1.34.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.6.1
numpy==1.26.4
openai==2.21.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==3.0.1
peft==0.18.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.10.0
tqdm==4.67.3
transformers==4.44.2
trec-car-tools==2.6
typer==0.24.0
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unlzw3==0.2.3
urllib3==2.6.3
uuid_utils==0.14.1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.22.0
zlib-state==0.1.10
zstandard==0.25.0
